requirements.txt
langchain
langchain-community
sentence-transformers 
langchain-huggingface
chromadb
pypdf
python-dotenv
langgraph
langchain_chroma
langchain-groq
streamlit
pandas
plotly
schedule
requests
redis